resolveNetworkID(): don't fail on pre-defined networks
When failing to resolve a network, but the network's name is (likely) a
pre-defined network, use the network's name as-is, instead of resolving
to a network ID.

Some APIs (such as the classic Swarm API) do not return pre-defined
networks, and therefore, deploying a service on one (such as `--network=host`)
would fail.

Pre-defined networks always exist on all nodes as local-scoped networks,
so there's no need to inspect them.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>