Some small fix-ups / improvements
1. Remove the "progress" line after copy is complete

After copying, the "progress" line was kept:

    docker cp foo:/usr/share/nginx bla
    Preparing to copy...

    docker cp foo:/usr/share/nginx bla
    Copying from container - 512B

    docker cp foo:/usr/share/nginx bla
    Copying from container - 4.608kB
    Successfully copied 4.608kB to /go/src/github.com/docker/cli/bla

The "Copying from container" line is a bit redundant after it's completed, so
removing the line afterwards:

    docker cp foo:/usr/share/nginx bla
    Preparing to copy...

    docker cp foo:/usr/share/nginx bla
    Copying from container - 512B

    docker cp foo:/usr/share/nginx bla
    Successfully copied 4.608kB to /go/src/github.com/docker/cli/bla

2. Print progress lines with a trailing newline

Without a newline, the cursor would be visible at the end of the progress
line, which was a bit distracting, as it would be jumping if the progress
line was not the same width for each refresh.

3. Swapped `fmt.Fprint` for `fmt.Fprintln` in some places

Same result, just prevents having to add `\n` in the strings, and it automatically
joins arguments with a space in between, so a slight bit cleaner.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>