WIP: debugging why it doesn't work
As a quickfix, update monitorContainerEvents to use a separate filter,
however this removes the feature that the Filter option was added for
in the first place.

We should either rewrite filters, or have some other way to use them.

The problem is also because of how labels are included in events, which
unfortunately merges them with other attributes (so not namespaced as
"label.label-foo").

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>